fixed display date, fixed medication list adapter
