+++ app/src/main/java/com/example/medstime/ui/medication/adapters/MedicationsListAdapter.kt
@@ -1,5 +1,6 @@
 package com.example.medstime.ui.medication.adapters
 
+import android.content.Context
 import android.os.Build
 import android.view.LayoutInflater
 import android.view.View
@@ -13,10 +14,10 @@ import java.time.LocalDateTime
 
 class MedicationsListAdapter(
     private val dataList: List<MedicationIntakeModel>,
-    private val medicationClick: (MedicationIntakeModel) -> Unit,
+    private val medicationClick: (MedicationIntakeModel, String) -> Unit,
+    private val context: Context,
 ) :
     RecyclerView.Adapter<MedicationsListAdapter.ViewHolder>() {
-
 
     inner class ViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
         val name: TextView = itemView.findViewById(R.id.medName)
@@ -34,11 +35,7 @@ class MedicationsListAdapter(
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
         val dataItem = dataList[position]
         holder.name.text = dataItem.name
-        holder.dosage.text = buildString {
-            append(dataItem.dosage.toString())
-            append(" ")
-            append(dataItem.dosageUnit)
-        }
+        holder.dosage.text = buildDosageString(dataItem)
         with(holder.timeMedication) {
             //если время приема не настало, текста не будет
             if (timeHasCome(dataItem)) {
@@ -55,7 +52,7 @@ class MedicationsListAdapter(
             }
         }
         holder.itemView.setOnClickListener {
-            medicationClick(dataItem)
+            medicationClick(dataItem, buildTimeAndDosageText(dataItem))
         }
 
     }
@@ -68,7 +65,6 @@ class MedicationsListAdapter(
     @RequiresApi(Build.VERSION_CODES.O)
     private fun timeHasCome(model: MedicationIntakeModel): Boolean {
         val currentDateTime = LocalDateTime.now()
-
         val intakeDateTime = LocalDateTime.of(
             currentDateTime.year,
             model.intakeDate.month,
@@ -76,14 +72,36 @@ class MedicationsListAdapter(
             model.intakeTime.hour,
             model.intakeTime.minute
         )
-
         return currentDateTime.isEqual(intakeDateTime) || currentDateTime.isAfter(intakeDateTime)
     }
 
-    private fun buildTimeString(time: MedicationIntakeModel.Time): String {
+
+    private fun buildTimeAndDosageText(medicationIntakeModel: MedicationIntakeModel): String {
+        val time = buildTimeString(medicationIntakeModel.intakeTime)
+        val dosage = buildDosageString(medicationIntakeModel)
+        return "$time $dosage"
+    }
+
+    private fun buildTimeString(time: MedicationIntakeModel.Time): String =
         with(time) {
             if (minute < 10) return "$hour:0$minute"
             else return "$hour:$minute"
         }
+
+    private fun buildDosageString(medicationIntakeModel: MedicationIntakeModel): String {
+        val commonText =
+            "${medicationIntakeModel.dosageUnit} ${medicationIntakeModel.intakeType.toRussianString()}"
+        return if (medicationIntakeModel.dosage == medicationIntakeModel.dosage.toInt().toDouble())
+            "${medicationIntakeModel.dosage.toInt()} $commonText"
+        else
+            "${medicationIntakeModel.dosage} $commonText"
     }
+
+    private fun MedicationIntakeModel.IntakeType.toRussianString() =
+        when (this) {
+            MedicationIntakeModel.IntakeType.AFTER_MEAL -> context.getString(R.string.after_meal)
+            MedicationIntakeModel.IntakeType.BEFORE_MEAL -> context.getString(R.string.before_meal)
+            MedicationIntakeModel.IntakeType.DURING_MEAL -> context.getString(R.string.during_meal)
+            MedicationIntakeModel.IntakeType.NONE -> context.getString(R.string.empty_string)
+        }
 }
